+++ src/renderer/components/ui/atoms/index.ts
@@ -1,6 +1,28 @@
 export { Card } from './Card';
-export { RaceTable, type HorseData as RaceTableHorseData } from './RaceTable';
+export { RaceTable } from './RaceTable';
+export { RaceTableResponsive } from './RaceTableResponsive';
 export { RaceCharts } from './RaceCharts';
 export { RecommendedBets, type RecommendedBet } from './RecommendedBets';
 export { ListTable, type HorseData as ListTableHorseData } from './ListTable';
 export { ListCharts } from './ListCharts';
+
+// 型定義のエクスポート
+export type {
+  HorseData,
+  SortField,
+  SortDirection,
+  RaceTableProps,
+} from './types/raceTable';
+
+// ユーティリティ関数のエクスポート
+export {
+  getEvaluationColor,
+  getGateColor,
+  sortHorseData,
+} from './utils/raceTableUtils';
+
+// サブコンポーネントのエクスポート
+export { HorseDetailModal } from './modals/HorseDetailModal';
+export { RaceTableHeader } from './table/RaceTableHeader';
+export { RaceTableBody } from './table/RaceTableBody';
+export { RaceTableMobileView } from './mobile/RaceTableMobileView';
